Force evaluation to be symmetric
With balanced=False, the result of evaluateComplete often varied by multiple pawns depending on the color argument. This adds a lot of noise, and is especially problematic with search extensions (which change the side to move at a leaf node for a whole subtree). Fixing the resulting blunders is worth the slow-down. (e.g. top engines can no longer give PyChess queen odds and still win.)